+++ Tools/RegionMapWidget.cpp
@@ -61,24 +61,25 @@ void RegionMapWidget::paintEvent(QPaintEvent* anEvent)
 	//	}
 	//}
 	for (qint64 x = l; x<= v.topRight().lon() / REGION_WIDTH; ++x) {
-			P.drawLine((x * REGION_WIDTH - v.bottomLeft().lon()) * sx, 0, (x * REGION_WIDTH - v.bottomLeft().lon()) * sx, height());
+            P.drawLine(int((x * REGION_WIDTH - v.bottomLeft().lon()) * sx), 0, int((x * REGION_WIDTH - v.bottomLeft().lon()) * sx), height());
 	}
 	for (qint64 y = t; y<= v.topRight().lat() / REGION_WIDTH; ++y) {
-			P.drawLine(0, height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy), width(), height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy));
+            P.drawLine(0, int(height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy)), width(), int(height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy)));
 	}
 	for (qint64 x = l; x<= v.topRight().lon() / REGION_WIDTH; ++x)
 		for (qint64 y = t; y<= v.topRight().lat() / REGION_WIDTH; ++y) {
 			quint32 rg = (y + NUM_REGIONS/2)*NUM_REGIONS + (x + NUM_REGIONS/2);
-			QRect rgRect = QRect((x * REGION_WIDTH - v.bottomLeft().lon()) * sx, 
-							height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy), REGION_WIDTH * sx, -REGION_WIDTH * sy);
+            QRect rgRect = QRect(int((x * REGION_WIDTH - v.bottomLeft().lon()) * sx),
+                            int(height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy)),
+                            int(REGION_WIDTH * sx), int(-REGION_WIDTH * sy));
 			//P.setClipRect(rgRect.adjusted(-1, -1, 1, 1));
 			if (ExistingRegions[rg])
-				P.drawText((x * REGION_WIDTH - v.bottomLeft().lon()) * sx, 
-							height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy),
+                P.drawText(int((x * REGION_WIDTH - v.bottomLeft().lon()) * sx),
+                            int(height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy)),
 							QString("%1 - %2").arg(QString::number(rg)).arg(DateRegions[rg].toString("yyyy-MM-dd")));
 			else
-				P.drawText((x * REGION_WIDTH - v.bottomLeft().lon()) * sx, 
-							height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy),
+                P.drawText(int((x * REGION_WIDTH - v.bottomLeft().lon()) * sx),
+                            int(height() - ((y * REGION_WIDTH - v.bottomLeft().lat()) * sy)),
 							QString("%1").arg(QString::number(rg)));
 
 			if (ExistingRegions[rg] && !DeleteRegions[rg]) {
@@ -113,7 +114,7 @@ void RegionMapWidget::mouseReleaseEvent(QMouseEvent* ev)
 		CoordBox v = CoordBox(Coord(R.y(), R.x()), Coord(R.y()+R.height(), R.x()+R.width()));
 		QPointF P = ev->pos();
 
-		Coord Pt(((height()-P.y()) / height() * v.latDiff()) + v.bottomLeft().lat(), (P.x() / width() * v.lonDiff()) + v.bottomLeft().lon());
+        Coord Pt(int(((height()-P.y()) / height() * v.latDiff()) + v.bottomLeft().lat()), int((P.x() / width() * v.lonDiff()) + v.bottomLeft().lon()));
 		int x = int(((qint64)Pt.lon()) / REGION_WIDTH);
 		x = (x < 0) ? x-1 :x;
 		int y = int(((qint64)Pt.lat()) / REGION_WIDTH);
